Add threshold to Prompt Scorer (#499)
* Add threshold to Prompt Scorer

* Add e2etests for prompt scorer
